Marked GetAvailableDatesQuery as legacy, and refactored code to use CostCalculator instead of GetTotalCostQuery
